staging ready for first large QA test

+++ install/table_gift_cards.sql
@@ -7,7 +7,7 @@ Staging Tables for the Magento SQL to Shopify API
 DROP TABLE IF EXISTS `shopify_gift_card`;
 CREATE TABLE `shopify_gift_card` (
   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
-  `shopify_id` int(20) DEFAULT NULL,
+  `shopify_id` bigint(20) DEFAULT NULL,
   `magento_id` int(20) DEFAULT NULL,
   `code` varchar(255) DEFAULT NULL,
   `initial_value` decimal(10,2) DEFAULT NULL,
