+++ install/tables_customers.sql
@@ -7,7 +7,7 @@ Staging Tables for the Magento SQL to Shopify API
 DROP TABLE IF EXISTS `shopify_customer`;
 CREATE TABLE `shopify_customer` (
   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
-  `shopify_id` int(20) DEFAULT NULL,
+  `shopify_id` bigint(20) DEFAULT NULL,
   `magento_id` int(20) DEFAULT NULL,
   `email` varchar(255) DEFAULT NULL,
   `first_name` varchar(255) DEFAULT NULL,
